+++ src/main/java/insane96mcp/iguanatweaksreborn/mixin/ThrowablePotionItemMixin.java
@@ -1,0 +1,19 @@
+package insane96mcp.iguanatweaksreborn.mixin;
+
+import com.llamalad7.mixinextras.injector.ModifyExpressionValue;
+import insane96mcp.iguanatweaksreborn.module.misc.Tweaks;
+import insane96mcp.insanelib.base.Feature;
+import net.minecraft.world.item.ThrowablePotionItem;
+import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.injection.At;
+
+@Mixin(ThrowablePotionItem.class)
+public class ThrowablePotionItemMixin {
+	@ModifyExpressionValue(method = "use", at = @At(value = "CONSTANT", args = "floatValue=0.5F"))
+	public float getUseDuration(float original) {
+		if (!Feature.isEnabled(Tweaks.class))
+			return original;
+
+		return Tweaks.splashPotionThrowStrength.floatValue();
+	}
+}