+++ changelog.md
@@ -1,6 +1,7 @@
 # Changelog
 
 ## Upcoming
+* Splash potions can now be thrown much farther
 * Desire paths no longer form if the player's sneaking
 * Fixed echo lantern dropping nothing
 * Fixed piercing damage ignored if the entity had regenerating absorption